pyproject.toml
[project]
name = "fbsurvivor"
version = "3.0"
description = "Run an NFL Survivor Pool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django",
    "django-extensions",
    "django-htmlmin",
    "pyjwt",
    "python-decouple",
    "requests",
    "gunicorn",
    "sentry-sdk[django]",
    "psycopg[c]",
]

[tool.uv]
dev-dependencies = [
    "django-debug-toolbar",
    "django-stubs",
    "django-types",
    "factory-boy",
    "ipython",
    "pyright[nodejs]",
    "pytest",
    "pytest-django",
]

[tool.ruff]
line-length = 99
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
exclude = [".git", ".pytest_cache", ".ruff_cache", "venv", ".venv"]

[tool.ruff.lint]
select = ["C", "E", "F", "W"]

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["**/__pycache__", "**/migrations", ".venv", "venv"]
ignore = ["**/migrations"]
typeCheckingMode = "basic"
reportMissingImports = "error"
reportUnusedImports = "error"
reportUnusedVariable = "error"
reportUnnecessaryTypeIgnoreComment = "error"
reportMissingTypeStubs = false
